RF | Delete JobPage and rename ProjectPage>JobPage
