+++ src/test/java/school/redrover/model/JobPage.java
@@ -2,9 +2,11 @@ package school.redrover.model;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedConditions;
 import school.redrover.model.base.BaseMainHeaderPage;
-import school.redrover.model.base.BaseModel;
-import school.redrover.runner.TestUtils;
+
+import java.time.Duration;
 
 public class JobPage extends BaseMainHeaderPage<JobPage> {
 
@@ -12,8 +14,59 @@ public class JobPage extends BaseMainHeaderPage<JobPage> {
         super(driver);
     }
 
-    public MainPage clickDashBoardButton() {
-        TestUtils.click(this, getDriver().findElement(By.xpath("//div[@id='breadcrumbBar']//a[text()='Dashboard']")));
-        return new  MainPage(getDriver());
+    public MainPage navigateToHomePageUsingJenkinsIcon() {
+        getWait5().until(ExpectedConditions.elementToBeClickable(getDriver()
+                .findElement(By.cssSelector("#jenkins-head-icon")))).click();
+        return new MainPage(getDriver());
+    }
+
+    public MainPage navigateToMainPageByBreadcrumbs() {
+        getWait2().until(ExpectedConditions.elementToBeClickable(getDriver()
+                        .findElement(By.xpath("//ol[@id='breadcrumbs']//li[1]")))).click();
+        return new MainPage(getDriver());
+    }
+
+    public WebElement getNameProject() {
+        return getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("#main-panel>h1")));
+    }
+
+    public WebElement getEnableForm(){
+        return getDriver().findElement(By.cssSelector("form#enable-project"));
+    }
+
+    public WebElement getDisableButton(){
+        return getDriver().findElement(By.xpath("//form[@id='disable-project']/button"));
+
+    }
+
+    public JobPage getDisableClick(){
+        getDriver().findElement(By.xpath("//form[@id='disable-project']/button")).click();
+        return this;
+    }
+
+    public RenamePage<JobPage> clickRename() {
+        getDriver().findElement(By.linkText("Rename")).click();
+        return new RenamePage<>(this);
+    }
+
+    public JobPage enableProject(){
+        getWait5().until(ExpectedConditions.elementToBeClickable(getDriver().
+                findElement(By.xpath("//form[@id='enable-project']/button")))).click();
+
+        return this;
+    }
+
+    public MainPage deleteProject(){
+        getDriver().findElement(By.xpath("//a/span[text()='Delete Multi-configuration project']/..")).click();
+        getDriver().switchTo().alert().accept();
+        getDriver().manage().timeouts().pageLoadTimeout(Duration.ofSeconds(2));
+        return new MainPage(getDriver());
+    }
+    public WebElement projectsHeadline() {
+        return getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.xpath(("//h1[contains(text(),'Project')]"))));
+    }
+
+    public String getProjectDescription() {
+        return getDriver().findElement(By.xpath("//div[@id='description']")).getText();
     }
 }
